refactor: removing duplicate testing implementations
