+++ packages/reactor/test/sync/channels/gql-req-channel/connection-state.test.ts
@@ -1,177 +1,16 @@
-import type { OperationContext } from "document-model";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
-import type { IOperationIndex } from "../../../../src/cache/operation-index-types.js";
-import type { ISyncCursorStorage } from "../../../../src/storage/interfaces.js";
+import { GqlRequestChannel } from "../../../../src/sync/channels/gql-req-channel.js";
+import type { ConnectionStateSnapshot } from "../../../../src/sync/types.js";
 import {
-  GqlRequestChannel,
-  type GqlChannelConfig,
-} from "../../../../src/sync/channels/gql-req-channel.js";
-import type { IPollTimer } from "../../../../src/sync/channels/poll-timer.js";
-import { SyncOperation } from "../../../../src/sync/sync-operation.js";
-import type {
-  ConnectionStateSnapshot,
-  RemoteFilter,
-} from "../../../../src/sync/types.js";
-import { createMockLogger } from "../../../factories.js";
-
-class ManualPollTimer implements IPollTimer {
-  private delegate: (() => Promise<void>) | undefined;
-  private running = false;
-
-  setDelegate(delegate: () => Promise<void>): void {
-    this.delegate = delegate;
-  }
-
-  start(): void {
-    this.running = true;
-  }
-
-  stop(): void {
-    this.running = false;
-  }
-
-  async tick(): Promise<void> {
-    if (this.running && this.delegate) {
-      await this.delegate();
-    }
-  }
-
-  isRunning(): boolean {
-    return this.running;
-  }
-}
-
-const TEST_FILTER: RemoteFilter = {
-  documentId: [],
-  scope: [],
-  branch: "main",
-};
-
-const createTestConfig = (
-  overrides: Partial<GqlChannelConfig> = {},
-): GqlChannelConfig => ({
-  url: "https://example.com/graphql",
-  collectionId: "test-collection",
-  filter: TEST_FILTER,
-  retryBaseDelayMs: 1000,
-  retryMaxDelayMs: 300000,
-  ...overrides,
-});
-
-const createMockCursorStorage = (): ISyncCursorStorage => ({
-  list: vi.fn().mockResolvedValue([]),
-  get: vi.fn().mockResolvedValue({
-    remoteName: "remote-1",
-    cursorType: "inbox",
-    cursorOrdinal: 0,
-  }),
-  upsert: vi.fn().mockResolvedValue(undefined),
-  remove: vi.fn().mockResolvedValue(undefined),
-});
-
-const createMockOperationIndex = (): IOperationIndex => ({
-  start: vi.fn(),
-  commit: vi.fn().mockResolvedValue([]),
-  find: vi
-    .fn()
-    .mockResolvedValue({ items: [], nextCursor: undefined, hasMore: false }),
-  get: vi
-    .fn()
-    .mockResolvedValue({ results: [], options: { cursor: "0", limit: 100 } }),
-  getSinceOrdinal: vi
-    .fn()
-    .mockResolvedValue({ items: [], nextCursor: undefined, hasMore: false }),
-  getLatestTimestampForCollection: vi.fn().mockResolvedValue(null),
-  getCollectionsForDocuments: vi.fn().mockResolvedValue({}),
-});
-
-const createMockOperationContext = (ordinal: number = 1): OperationContext => ({
-  documentId: "doc-1",
-  documentType: "test/document",
-  scope: "public",
-  branch: "main",
-  ordinal,
-});
-
-const createMockSyncOperation = (
-  id: string,
-  remoteName: string,
-  ordinal: number = 0,
-): SyncOperation => {
-  return new SyncOperation(
-    id,
-    "",
-    [],
-    remoteName,
-    "doc-1",
-    ["public"],
-    "main",
-    [
-      {
-        operation: {
-          index: 0,
-          skip: 0,
-          id: "op-1",
-          timestampUtcMs: new Date().toISOString(),
-          hash: "hash-1",
-          action: {
-            type: "TEST_OP",
-            id: "action-1",
-            scope: "public",
-            timestampUtcMs: new Date().toISOString(),
-            input: {},
-          },
-        },
-        context: createMockOperationContext(ordinal),
-      },
-    ],
-  );
-};
-
-function createMockFetch(
-  handler: (body: { query: string }) => {
-    ok: boolean;
-    json: () => Promise<unknown>;
-  },
-) {
-  return vi.fn().mockImplementation((_url: string, options: RequestInit) => {
-    const body = JSON.parse(options.body as string) as { query: string };
-    return Promise.resolve(handler(body));
-  });
-}
-
-function successFetch() {
-  return createMockFetch((body) => {
-    if (body.query.includes("touchChannel")) {
-      return {
-        ok: true,
-        json: () =>
-          Promise.resolve({
-            data: { touchChannel: { success: true, ackOrdinal: 0 } },
-          }),
-      };
-    }
-    if (body.query.includes("pushSyncEnvelopes")) {
-      return {
-        ok: true,
-        json: () => Promise.resolve({ data: { pushSyncEnvelopes: true } }),
-      };
-    }
-    return {
-      ok: true,
-      json: () =>
-        Promise.resolve({
-          data: {
-            pollSyncEnvelopes: {
-              envelopes: [],
-              ackOrdinal: 0,
-              deadLetters: [],
-            },
-          },
-        }),
-    };
-  });
-}
+  ManualPollTimer,
+  createMockCursorStorage,
+  createMockFetch,
+  createMockLogger,
+  createMockOperationIndex,
+  createMockSyncOperation,
+  createTestConfig,
+  successFetch,
+} from "./test-helpers.js";
 
 describe("GqlRequestChannel Connection State", () => {
   let originalFetch: typeof global.fetch;
